updates user specific routes to check for session data

+++ flask_app/flask06.py
@@ -38,9 +38,9 @@ notes =     {1: {'title': 'First note', 'text': 'This is my first note', 'date':
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='jespin12@uncc.edu').one()
-
-    return render_template ("index.html" , user = a_user)
+    if session.get('user'):
+        return render_template ("index.html" , user = session['user'])
+    return render_template("index.html")
 
 @app.route('/notes')
 def get_notes():
@@ -62,53 +62,64 @@ def get_note(note_id):
     return render_template ("note.html", note=my_note, user=a_user)
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
-    if request.method == 'POST' :
-        title = request.form['title']
+    if session.get('user'):
 
-        text = request.form['noteText']
+        if request.method == 'POST' :
+            title = request.form['title']
 
-        from datetime import date
-        today = date.today()
+            text = request.form['noteText']
 
-        today = today.strftime("%m-%d-%Y")
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
+            from datetime import date
+            today = date.today()
 
-        return redirect(url_for('get_notes'))
+            today = today.strftime("%m-%d-%Y")
+            new_record = Note(title, text, today)
+            db.session.add(new_record)
+            db.session.commit()
 
-    else: 
-        a_user = db.session.query(User).filter_by(email='jespin12@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+            return redirect(url_for('get_notes'))
+
+        else: 
+            return render_template('new.html', user=session['user'])
+    else:
+        return render_template('new.html', user=session['user'])
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    if request.method == 'POST':
-        title = request.form['title']
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
 
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
 
-        note.title = title
-        note.text = text
+            note.title = title
+            note.text = text
         
-        db.session.add(note)
-        db.session.commit()
+            db.session.add(note)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
-    else: 
-        a_user = db.session.query(User).filter_by(email='jespin12@uncc.edu').one()
+            return redirect(url_for('get_notes'))
+        else: 
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
 
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        return render_template('new.html', note=my_note, user = session['user'])
 
-        return render_template('new.html', note=my_note, user = a_user)
+    else:
+        return redirect(url_for('login'))
 
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
+    if session.get('user'):
 
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
+
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect(url_for('login'))
+        
 @app.route('/register', methods=['POST','GET'])
 def register():
     form = RegisterForm()
